refactor(example-data): Separate RichText from Convenience
This provides slightly better idea, which of the RichText
factory methods are convenience and which are based on
plain CoreMedia RichText 1.0.

Kept `RichText` as integrator of both APIs.
